Modified automation case for CSA
